+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.0] · Not Yet Released
+## [0.7.0] · 2020-05-29
 
 ### Changed
 - `toBe` shows prints difference between received and expected values (like toEqual) if it received non-reference types.